fix(toaster): width style issue
