+++ src/constants/navigateOption.ts
@@ -1,0 +1,1 @@
+export const RESET_SCREEN = 'RESET_SCREEN';
